Add files via upload
Concatenate_File.m is a MATLAB script that searches within a specific file pathway to view all .set files (EEGLAB) with a specific string within the file name; in this case, ICArem. It then merges all data together and finds the common sampling rate (this will be changed in later updates) and re-samples all sets to the common.